Cache/requirements.txt
# Chat Service Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
redis==5.0.1
PyYAML==6.0.1
python-dotenv==1.0.0